Create Movie Catalog
You are developing a program to manage a list of movies in a cinema catalog using a double-linked list. Each node in the list represents a movie, and each movie has a title and a release year. The program should allow users to add movies to the cinema's catalog and print the details of all the movies.

Implement a function to create a double-linked list node for a movie. The node should contain fields for the title and release year.
Create a function to insert a new movie at the end of the double-linked list.
Implement a function to print the details of all the movies in the cinema's catalog.


Write a C program that incorporates these functions to manage the cinema's movie catalog.



Input Format:

The first line of input contains an integer N representing the number of movies in the catalog.
The next N lines contain details for each movie in the format: "Title Release_Year".
Output Format:
Print the details of all the movies in the catalog. (Refer to the visible test cases for better understanding the format)
Note: The name of the movie should not have any spaces in it