Create  Reverse a String
You are building a feature for a messaging app that allows users to send encrypted messages. To decrypt them, you need to reverse the encrypted message. Implement a function reverseMessage() that takes a message (string) as input and returns the decrypted message. For this, you need to use the stack data structure.



Input format:

The input is a string that represents the encrypted message.



Output format:

The output is a reversed string that represents the decrypted message.



Note:

The implementation of the stack data structure and main function is already given to you, you just need to fill the code in the reverseMessage() function by using the implemented stack.

Test case1:
program	
margorp	
Test case 2:
codetantra@1234	
4321@artnatedoc