Create Palindromic Strings
Imagine you're in a classroom where the teacher is talking about special words that look the same even if you read them backward. These cool words are called palindromes, and today's lesson is all about them!



Our friend Ayan is really excited to learn more about these special words. For that he wants to write a program that should be like a word detective, figuring out if a word is a palindrome or not. Could you help him out by creating a C program?



Remember you must use stacks in order to determine whether the given words are palindrome or not.



Input format:

Prompts the user to enter the string


Output format:

The output returns whether it is palindrome or not palindrome based on the input given.


Note:

The implementation of the stack data structure and main function is already given to you, you just need to fill the code in the isPalindrome() function by using the implemented stack.


Test case 1
madam	
palindrome⏎	
Test case 2
malayalam	
palindrome⏎	
Test case 3
GemGem	
not·palindrome⏎